+++ packages/preact_signals/example/web/main.dart
@@ -4,21 +4,20 @@ import 'package:preact_signals/preact_signals.dart';
 typedef Task = ({String title, bool completed});
 
 void main() {
-  // Application logic
   final todoForm = document.getElementById("todoForm")!;
   final todoInput = document.getElementById("todoInput") as InputElement;
   final todoList = document.getElementById("todoList")!;
   final taskFilter = document.getElementById("taskFilter")!;
   final taskCounter = document.getElementById("taskCounter")!;
 
-  final tasks = signal<List<Task>>([]);
+  final tasks = <Task>[].toSignal();
   final filter = signal("all");
 
   final filteredTasks = computed(() {
     final currentFilter = filter.value;
-    final currentTasks = tasks.value;
+    final currentTasks = tasks;
     if (currentFilter == "all") {
-      return currentTasks;
+      return currentTasks.toList();
     } else if (currentFilter == "active") {
       return currentTasks.where((task) => !task.completed).toList();
     } else {
@@ -27,11 +26,11 @@ void main() {
   });
 
   final taskCount = computed(() {
-    return tasks.value.length;
+    return tasks.length;
   });
 
   final activeTaskCount = computed(() {
-    return tasks.value.where((task) => !task.completed).length;
+    return tasks.where((task) => !task.completed).length;
   });
 
   final completedTaskCount = computed(() {
@@ -46,7 +45,7 @@ void main() {
     final taskTitle = todoInput.value?.trim();
     if (taskTitle != null) {
       final newTask = (title: taskTitle, completed: false);
-      tasks.value = [...tasks.value, newTask];
+      tasks.add(newTask);
       todoInput.value = "";
     }
   });
@@ -67,12 +66,10 @@ void main() {
       checkbox.type = "checkbox";
       checkbox.checked = task.completed;
       checkbox.addEventListener("change", (e) {
-        final current = tasks.value;
-        current[index] = (
-          title: current[index].title,
+        tasks[index] = (
+          title: tasks[index].title,
           completed: checkbox.checked ?? false,
         );
-        tasks.value = [...current];
       });
       label.append(checkbox);
       label.append(Text(" ${task.title}"));
@@ -83,9 +80,9 @@ void main() {
 
   effect(() {
     taskCounter.text = '''
-        Total: ${taskCount.value}, 
-        Active: ${activeTaskCount.value}, 
-        Completed: ${completedTaskCount.value}
-        ''';
+    Total: ${taskCount.value}, 
+    Active: ${activeTaskCount.value}, 
+    Completed: ${completedTaskCount.value}
+    ''';
   });
 }