+++ packages/preact_signals/lib/preact_signals.dart
@@ -7,12 +7,21 @@ export 'src/signals.dart'
         effect,
         batch,
         untracked,
+        Effect,
+        Computed,
         Signal,
         MutableSignal,
-        ReadonlySignal,
-        initPreactSignalsDevTools;
+        ReadonlySignal;
 export 'src/extensions/stream.dart';
 export 'src/extensions/future.dart';
+export 'src/extensions/list.dart';
+export 'src/extensions/map.dart';
+export 'src/extensions/set.dart';
+export 'src/extensions/iterable.dart';
 export 'src/signal_state.dart';
-export 'src/future_signal.dart';
-export 'src/stream_signal.dart';
+export 'src/future_signal.dart' show FutureSignal, futureSignal;
+export 'src/stream_signal.dart' show StreamSignal, streamSignal;
+export 'src/list_signal.dart' show ListSignal, listSignal;
+export 'src/map_signal.dart' show MapSignal, mapSignal;
+export 'src/set_signal.dart' show SetSignal, setSignal;
+export 'src/iterable_signal.dart' show IterableSignal, iterableSignal;