+++ packages/preact_signals/lib/src/extensions/list.dart
@@ -1,0 +1,9 @@
+import 'package:preact_signals/src/list_signal.dart';
+
+/// Extension on future to provide helpful methods for signals
+extension SignalListUtils<T> on List<T> {
+  /// Convert an existing list to [ListSignal]
+  ListSignal<T> toSignal() {
+    return ListSignal(this);
+  }
+}