+++ packages/preact_signals/lib/src/signals.dart
@@ -400,6 +400,19 @@ class Signal<T> implements MutableSignal<T> {
     return true;
   }
 
+  @override
+  bool operator ==(Object other) {
+    return other is Signal<T> && value == other.value;
+  }
+
+  @override
+  int get hashCode {
+    return Object.hashAll([
+      globalId.hashCode,
+      value.hashCode,
+    ]);
+  }
+
   static void __subscribe(ReadonlySignal signal, Node node) {
     if (signal._targets != node && node._prevTarget == null) {
       node._nextTarget = signal._targets;
@@ -487,22 +500,30 @@ class Signal<T> implements MutableSignal<T> {
     }
 
     if (val != this._value) {
-      if (batchIteration > 100) {
-        cycleDetected();
-      }
+      _updateValue(val);
+    }
+  }
 
-      this._value = val;
-      this._version++;
-      globalVersion++;
+  void forceUpdate(T val) {
+    _updateValue(val);
+  }
 
-      startBatch();
-      try {
-        for (var node = _targets; node != null; node = node._nextTarget) {
-          node._target._notify();
-        }
-      } finally {
-        endBatch();
+  void _updateValue(T val) {
+    if (batchIteration > 100) {
+      cycleDetected();
+    }
+
+    this._value = val;
+    this._version++;
+    globalVersion++;
+
+    startBatch();
+    try {
+      for (var node = _targets; node != null; node = node._nextTarget) {
+        node._target._notify();
       }
+    } finally {
+      endBatch();
     }
   }
 
@@ -670,6 +691,19 @@ class Computed<T> implements Listenable, ReadonlySignal<T> {
 
   bool _initialized = false;
   late T _value;
+
+  @override
+  bool operator ==(Object other) {
+    return other is Signal<T> && value == other.value;
+  }
+
+  @override
+  int get hashCode {
+    return Object.hashAll([
+      globalId.hashCode,
+      value.hashCode,
+    ]);
+  }
 
   Computed(ComputedCallback<T> compute, {this.debugLabel})
       : _compute = compute,
@@ -933,6 +967,16 @@ class Effect implements Listenable {
 
   @override
   int _flags;
+
+  @override
+  bool operator ==(Object other) {
+    return other is Effect && globalId == other.globalId;
+  }
+
+  @override
+  int get hashCode {
+    return globalId.hashCode;
+  }
 
   Effect(EffectCallback compute, {this.debugLabel})
       : _flags = TRACKING,